refactor(ui): flatten `userinfo/` and `workflow-event-bindings/` dirs
- same as previous commits
