+++ ui/src/app/widgets/widgets.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
+
 import {ErrorNotice} from '../shared/components/error-notice';
 import {WorkflowGraph} from './workflow-graph';
 import {WorkflowStatusBadge} from './workflow-status-badge';
