+++ ui/src/app/widgets/workflow-graph.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import {useEffect, useState} from 'react';
 import {RouteComponentProps} from 'react-router';
+
 import {ArtifactRepositoryRefStatus, NodeStatus} from '../../models';
 import {uiUrl} from '../shared/base';
 import {ErrorNotice} from '../shared/components/error-notice';
