+++ ui/src/app/workflow-event-bindings/index.ts
@@ -1,4 +1,4 @@
-import {WorkflowEventBindingsContainer} from './components/workflow-event-bindings-container';
+import {WorkflowEventBindingsContainer} from './workflow-event-bindings-container';
 
 export default {
     component: WorkflowEventBindingsContainer
