+++ ui/src/app/workflow-event-bindings/workflow-event-bindings-container.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import {Route, RouteComponentProps, Switch} from 'react-router';
-import {WorkflowEventBindings} from './workflow-event-bindings/workflow-event-bindings';
+
+import {WorkflowEventBindings} from './workflow-event-bindings';
 
 export const WorkflowEventBindingsContainer = (props: RouteComponentProps<any>) => (
     <Switch>
