+++ ui/src/app/workflow-event-bindings/workflow-event-bindings.tsx
@@ -2,23 +2,24 @@ import {Page, SlidingPanel} from 'argo-ui';
 import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
 import {RouteComponentProps} from 'react-router-dom';
-import {WorkflowEventBinding} from '../../../../models';
-import {absoluteUrl, uiUrl} from '../../../shared/base';
-import {ErrorNotice} from '../../../shared/components/error-notice';
-import {InfoIcon} from '../../../shared/components/fa-icons';
-import {GraphPanel} from '../../../shared/components/graph/graph-panel';
-import {Graph} from '../../../shared/components/graph/types';
-import {Loading} from '../../../shared/components/loading';
-import {NamespaceFilter} from '../../../shared/components/namespace-filter';
-import {ResourceEditor} from '../../../shared/components/resource-editor/resource-editor';
-import {useCollectEvent} from '../../../shared/components/use-collect-event';
-import {ZeroState} from '../../../shared/components/zero-state';
-import {Context} from '../../../shared/context';
-import {Footnote} from '../../../shared/footnote';
-import {historyUrl} from '../../../shared/history';
-import {services} from '../../../shared/services';
-import {useQueryParams} from '../../../shared/use-query-params';
-import {Utils} from '../../../shared/utils';
+
+import {WorkflowEventBinding} from '../../models';
+import {absoluteUrl, uiUrl} from '../shared/base';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {InfoIcon} from '../shared/components/fa-icons';
+import {GraphPanel} from '../shared/components/graph/graph-panel';
+import {Graph} from '../shared/components/graph/types';
+import {Loading} from '../shared/components/loading';
+import {NamespaceFilter} from '../shared/components/namespace-filter';
+import {ResourceEditor} from '../shared/components/resource-editor/resource-editor';
+import {useCollectEvent} from '../shared/use-collect-event';
+import {ZeroState} from '../shared/components/zero-state';
+import {Context} from '../shared/context';
+import {Footnote} from '../shared/footnote';
+import {historyUrl} from '../shared/history';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {Utils} from '../shared/utils';
 import {ID} from './id';
 
 const introductionText = (
